Move launchUrl() to its own module
This way it can be shared.
